requirements.txt
requests>=2.25.1
PyYAML>=6.0
pathlib2>=2.3.0; python_version < '3.4'
